Fix script paths, add missing file and tests
