+++ index.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Offline Soil Sample Analyzer</title>
-  <script type="module" src="app.js"></script>
+  <script type="module" src="app.js" defer></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/PapaParse/5.3.2/papaparse.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/3.4.120/pdf.min.js"></script>
 </head>
